--- src/classes/screens/menu-actions.js
@@ -1,4 +1,0 @@
-const hideAllMenus = 'hide'
-const startGame = 'start'
-
-export { hideAllMenus, startGame }
